chore: split withBrowser into withBrowser and withMcpContext
