+++ tests/tools/input.test.ts
@@ -20,14 +20,14 @@ import {
 } from '../../src/tools/input.js';
 import {parseKey} from '../../src/utils/keyboard.js';
 import {serverHooks} from '../server.js';
-import {html, withBrowser} from '../utils.js';
+import {html, withMcpContext} from '../utils.js';
 
 describe('input', () => {
   const server = serverHooks();
 
   describe('click', () => {
     it('clicks', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.setContent(
           html`<button onclick="this.innerText = 'clicked';">test</button>`,
@@ -51,7 +51,7 @@ describe('input', () => {
       });
     });
     it('double clicks', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.setContent(
           html`<button ondblclick="this.innerText = 'dblclicked';"
@@ -89,7 +89,7 @@ describe('input', () => {
         res.end();
       });
 
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.goto(server.getRoute('/link'));
         await context.createTextSnapshot();
@@ -131,7 +131,7 @@ describe('input', () => {
           </script>
         `,
       );
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.goto(server.getRoute('/unstable'));
         await context.createTextSnapshot();
@@ -158,7 +158,7 @@ describe('input', () => {
 
   describe('hover', () => {
     it('hovers', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.setContent(
           html`<button onmouseover="this.innerText = 'hovered';">test</button>`,
@@ -185,7 +185,7 @@ describe('input', () => {
 
   describe('fill', () => {
     it('fills out an input', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.setContent(html`<input />`);
         await context.createTextSnapshot();
@@ -209,7 +209,7 @@ describe('input', () => {
     });
 
     it('fills out a select by text', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.setContent(
           html`<select
@@ -243,7 +243,7 @@ describe('input', () => {
 
   describe('drags', () => {
     it('drags one element onto another', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.setContent(
           html`<div
@@ -297,7 +297,7 @@ describe('input', () => {
 
   describe('fill form', () => {
     it('successfully fills out the form', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.setContent(
           html`<form>
@@ -361,7 +361,7 @@ describe('input', () => {
       const testFilePath = path.join(process.cwd(), 'test.txt');
       await fs.writeFile(testFilePath, 'test file content');
 
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.setContent(
           html`<form>
@@ -396,7 +396,7 @@ describe('input', () => {
       const testFilePath = path.join(process.cwd(), 'test.txt');
       await fs.writeFile(testFilePath, 'test file content');
 
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.setContent(
           html`<button id="file-chooser-button">Upload file</button>
@@ -443,7 +443,7 @@ describe('input', () => {
       const testFilePath = path.join(process.cwd(), 'test.txt');
       await fs.writeFile(testFilePath, 'test file content');
 
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.setContent(html`<div>Not a file input</div>`);
         await context.createTextSnapshot();
@@ -502,7 +502,7 @@ describe('input', () => {
     });
 
     it('processes press_key', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.setContent(
           html`<script>
