+++ tests/tools/network.test.ts
@@ -12,13 +12,13 @@ import {
   listNetworkRequests,
 } from '../../src/tools/network.js';
 import {serverHooks} from '../server.js';
-import {html, withBrowser, stabilizeResponseOutput} from '../utils.js';
+import {html, withMcpContext, stabilizeResponseOutput} from '../utils.js';
 
 describe('network', () => {
   const server = serverHooks();
   describe('network_list_requests', () => {
     it('list requests', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await listNetworkRequests.handler({params: {}}, response, context);
         assert.ok(response.includeNetworkRequests);
         assert.strictEqual(response.networkRequestsPageIdx, undefined);
@@ -30,7 +30,7 @@ describe('network', () => {
       server.addHtmlRoute('/two', html`<main>Second</main>`);
       server.addHtmlRoute('/three', html`<main>Third</main>`);
 
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await context.setUpNetworkCollectorForTesting();
         const page = context.getSelectedPage();
         await page.goto(server.getRoute('/one'));
@@ -53,7 +53,7 @@ describe('network', () => {
       server.addHtmlRoute('/two', html`<main>Second</main>`);
       server.addHtmlRoute('/three', html`<main>Third</main>`);
 
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await context.setUpNetworkCollectorForTesting();
         const page = context.getSelectedPage();
         await page.goto(server.getRoute('/one'));
@@ -93,7 +93,7 @@ describe('network', () => {
         html`<main>I was redirected 2 times</main>`,
       );
 
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await context.setUpNetworkCollectorForTesting();
         const page = context.getSelectedPage();
         await page.goto(server.getRoute('/redirect'));
@@ -113,7 +113,7 @@ describe('network', () => {
   });
   describe('network_get_request', () => {
     it('attaches request', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.goto('data:text/html,<div>Hello MCP</div>');
         await getNetworkRequest.handler(
@@ -126,7 +126,7 @@ describe('network', () => {
       });
     });
     it('should not add the request list', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.goto('data:text/html,<div>Hello MCP</div>');
         await getNetworkRequest.handler(
@@ -142,7 +142,7 @@ describe('network', () => {
       server.addHtmlRoute('/two', html`<main>Second</main>`);
       server.addHtmlRoute('/three', html`<main>Third</main>`);
 
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await context.setUpNetworkCollectorForTesting();
         const page = context.getSelectedPage();
         await page.goto(server.getRoute('/one'));
