+++ tests/tools/pages.test.ts
@@ -18,12 +18,12 @@ import {
   resizePage,
   handleDialog,
 } from '../../src/tools/pages.js';
-import {withBrowser} from '../utils.js';
+import {withMcpContext} from '../utils.js';
 
 describe('pages', () => {
   describe('list_pages', () => {
     it('list pages', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await listPages.handler({params: {}}, response, context);
         assert.ok(response.includePages);
       });
@@ -31,7 +31,7 @@ describe('pages', () => {
   });
   describe('new_page', () => {
     it('create a page', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         assert.strictEqual(context.getPageByIdx(0), context.getSelectedPage());
         await newPage.handler(
           {params: {url: 'about:blank'}},
@@ -45,7 +45,7 @@ describe('pages', () => {
   });
   describe('close_page', () => {
     it('closes a page', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = await context.newPage();
         assert.strictEqual(context.getPageByIdx(1), context.getSelectedPage());
         assert.strictEqual(context.getPageByIdx(1), page);
@@ -55,7 +55,7 @@ describe('pages', () => {
       });
     });
     it('cannot close the last page', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await closePage.handler({params: {pageIdx: 0}}, response, context);
         assert.deepStrictEqual(
@@ -69,7 +69,7 @@ describe('pages', () => {
   });
   describe('select_page', () => {
     it('selects a page', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await context.newPage();
         assert.strictEqual(context.getPageByIdx(1), context.getSelectedPage());
         await selectPage.handler({params: {pageIdx: 0}}, response, context);
@@ -80,7 +80,7 @@ describe('pages', () => {
   });
   describe('navigate_page', () => {
     it('navigates to correct page', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await navigatePage.handler(
           {params: {url: 'data:text/html,<div>Hello MCP</div>'}},
           response,
@@ -96,7 +96,7 @@ describe('pages', () => {
     });
 
     it('throws an error if the page was closed not by the MCP server', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = await context.newPage();
         assert.strictEqual(context.getPageByIdx(1), context.getSelectedPage());
         assert.strictEqual(context.getPageByIdx(1), page);
@@ -119,7 +119,7 @@ describe('pages', () => {
       });
     });
     it('go back', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.goto('data:text/html,<div>Hello MCP</div>');
         await navigatePage.handler({params: {type: 'back'}}, response, context);
@@ -132,7 +132,7 @@ describe('pages', () => {
       });
     });
     it('go forward', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.goto('data:text/html,<div>Hello MCP</div>');
         await page.goBack();
@@ -150,7 +150,7 @@ describe('pages', () => {
       });
     });
     it('reload', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.goto('data:text/html,<div>Hello MCP</div>');
         await navigatePage.handler(
@@ -167,7 +167,7 @@ describe('pages', () => {
       });
     });
     it('go forward with error', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await navigatePage.handler(
           {params: {type: 'forward'}},
           response,
@@ -183,7 +183,7 @@ describe('pages', () => {
       });
     });
     it('go back with error', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await navigatePage.handler({params: {type: 'back'}}, response, context);
 
         assert.ok(
@@ -197,7 +197,7 @@ describe('pages', () => {
   });
   describe('resize', () => {
     it('create a page', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         const resizePromise = page.evaluate(() => {
           return new Promise(resolve => {
@@ -220,7 +220,7 @@ describe('pages', () => {
 
   describe('dialogs', () => {
     it('can accept dialogs', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         const dialogPromise = new Promise<void>(resolve => {
           page.on('dialog', () => {
@@ -248,7 +248,7 @@ describe('pages', () => {
       });
     });
     it('can dismiss dialogs', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         const dialogPromise = new Promise<void>(resolve => {
           page.on('dialog', () => {
@@ -276,7 +276,7 @@ describe('pages', () => {
       });
     });
     it('can dismiss already dismissed dialog dialogs', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         const dialogPromise = new Promise<Dialog>(resolve => {
           page.on('dialog', dialog => {
