+++ tests/tools/performance.test.ts
@@ -20,7 +20,7 @@ import {
   traceResultIsSuccess,
 } from '../../src/trace-processing/parse.js';
 import {loadTraceAsBuffer} from '../trace-processing/fixtures/load.js';
-import {withBrowser} from '../utils.js';
+import {withMcpContext} from '../utils.js';
 
 describe('performance', () => {
   afterEach(() => {
@@ -29,7 +29,7 @@ describe('performance', () => {
 
   describe('performance_start_trace', () => {
     it('starts a trace recording', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         context.setIsRunningPerformanceTrace(false);
         const selectedPage = context.getSelectedPage();
         const startTracingStub = sinon.stub(selectedPage.tracing, 'start');
@@ -49,7 +49,7 @@ describe('performance', () => {
     });
 
     it('can navigate to about:blank and record a page reload', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const selectedPage = context.getSelectedPage();
         sinon.stub(selectedPage, 'url').callsFake(() => 'https://www.test.com');
         const gotoStub = sinon.stub(selectedPage, 'goto');
@@ -78,7 +78,7 @@ describe('performance', () => {
     it('can autostop and store a recording', async () => {
       const rawData = loadTraceAsBuffer('basic-trace.json.gz');
 
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const selectedPage = context.getSelectedPage();
         sinon.stub(selectedPage, 'url').callsFake(() => 'https://www.test.com');
         sinon.stub(selectedPage, 'goto').callsFake(() => Promise.resolve(null));
@@ -121,7 +121,7 @@ describe('performance', () => {
     });
 
     it('errors if a recording is already active', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         context.setIsRunningPerformanceTrace(true);
         const selectedPage = context.getSelectedPage();
         const startTracingStub = sinon.stub(selectedPage.tracing, 'start');
@@ -152,7 +152,7 @@ describe('performance', () => {
 
     it('returns the information on the insight', async t => {
       const trace = await parseTrace('web-dev-with-commit.json.gz');
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         context.storeTraceRecording(trace);
         context.setIsRunningPerformanceTrace(false);
 
@@ -173,7 +173,7 @@ describe('performance', () => {
 
     it('returns an error if the insight does not exist', async () => {
       const trace = await parseTrace('web-dev-with-commit.json.gz');
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         context.storeTraceRecording(trace);
         context.setIsRunningPerformanceTrace(false);
 
@@ -196,7 +196,7 @@ describe('performance', () => {
     });
 
     it('returns an error if no trace has been recorded', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await analyzeInsight.handler(
           {
             params: {
@@ -220,7 +220,7 @@ describe('performance', () => {
 
   describe('performance_stop_trace', () => {
     it('does nothing if the trace is not running and does not error', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         context.setIsRunningPerformanceTrace(false);
         const selectedPage = context.getSelectedPage();
         const stopTracingStub = sinon.stub(selectedPage.tracing, 'stop');
@@ -232,7 +232,7 @@ describe('performance', () => {
 
     it('will stop the trace and return trace info when a trace is running', async () => {
       const rawData = loadTraceAsBuffer('basic-trace.json.gz');
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         context.setIsRunningPerformanceTrace(true);
         const selectedPage = context.getSelectedPage();
         const stopTracingStub = sinon
@@ -252,7 +252,7 @@ describe('performance', () => {
     });
 
     it('returns an error message if parsing the trace buffer fails', async t => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         context.setIsRunningPerformanceTrace(true);
         const selectedPage = context.getSelectedPage();
         sinon
@@ -265,7 +265,7 @@ describe('performance', () => {
 
     it('returns the high level summary of the performance trace', async t => {
       const rawData = loadTraceAsBuffer('web-dev-with-commit.json.gz');
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         context.setIsRunningPerformanceTrace(true);
         const selectedPage = context.getSelectedPage();
         sinon.stub(selectedPage.tracing, 'stop').callsFake(async () => {
