+++ tests/tools/screenshot.test.ts
@@ -12,12 +12,12 @@ import {describe, it} from 'node:test';
 
 import {screenshot} from '../../src/tools/screenshot.js';
 import {screenshots} from '../snapshot.js';
-import {html, withBrowser} from '../utils.js';
+import {html, withMcpContext} from '../utils.js';
 
 describe('screenshot', () => {
   describe('browser_take_screenshot', () => {
     it('with default options', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const fixture = screenshots.basic;
         const page = context.getSelectedPage();
         await page.setContent(fixture.html);
@@ -32,7 +32,7 @@ describe('screenshot', () => {
       });
     });
     it('ignores quality', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const fixture = screenshots.basic;
         const page = context.getSelectedPage();
         await page.setContent(fixture.html);
@@ -51,7 +51,7 @@ describe('screenshot', () => {
       });
     });
     it('with jpeg', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await screenshot.handler({params: {format: 'jpeg'}}, response, context);
 
         assert.equal(response.images.length, 1);
@@ -63,7 +63,7 @@ describe('screenshot', () => {
       });
     });
     it('with webp', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await screenshot.handler({params: {format: 'webp'}}, response, context);
 
         assert.equal(response.images.length, 1);
@@ -75,7 +75,7 @@ describe('screenshot', () => {
       });
     });
     it('with full page', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const fixture = screenshots.viewportOverflow;
         const page = context.getSelectedPage();
         await page.setContent(fixture.html);
@@ -95,7 +95,7 @@ describe('screenshot', () => {
     });
 
     it('with full page resulting in a large screenshot', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
 
         await page.setContent(
@@ -129,7 +129,7 @@ describe('screenshot', () => {
     });
 
     it('with element uid', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const fixture = screenshots.button;
 
         const page = context.getSelectedPage();
@@ -156,7 +156,7 @@ describe('screenshot', () => {
     });
 
     it('with filePath', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const filePath = join(tmpdir(), 'test-screenshot.png');
         try {
           const fixture = screenshots.basic;
@@ -198,7 +198,7 @@ describe('screenshot', () => {
         await chmod(filePath, 0o400);
 
         try {
-          await withBrowser(async (response, context) => {
+          await withMcpContext(async (response, context) => {
             const fixture = screenshots.basic;
             const page = context.getSelectedPage();
             await page.setContent(fixture.html);
@@ -222,7 +222,7 @@ describe('screenshot', () => {
         const filePath = join(dir, 'test-screenshot.png');
 
         try {
-          await withBrowser(async (response, context) => {
+          await withMcpContext(async (response, context) => {
             const fixture = screenshots.basic;
             const page = context.getSelectedPage();
             await page.setContent(fixture.html);
@@ -242,7 +242,7 @@ describe('screenshot', () => {
     });
 
     it('with malformed filePath', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         // Use a platform-specific invalid character.
         // On Windows, characters like '<', '>', ':', '"', '/', '\', '|', '?', '*' are invalid.
         // On POSIX, the null byte is invalid.
