+++ tests/tools/script.test.ts
@@ -9,14 +9,14 @@ import {describe, it} from 'node:test';
 
 import {evaluateScript} from '../../src/tools/script.js';
 import {serverHooks} from '../server.js';
-import {html, withBrowser} from '../utils.js';
+import {html, withMcpContext} from '../utils.js';
 
 describe('script', () => {
   const server = serverHooks();
 
   describe('browser_evaluate_script', () => {
     it('evaluates', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await evaluateScript.handler(
           {params: {function: String(() => 2 * 5)}},
           response,
@@ -27,7 +27,7 @@ describe('script', () => {
       });
     });
     it('runs in selected page', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await evaluateScript.handler(
           {params: {function: String(() => document.title)}},
           response,
@@ -57,7 +57,7 @@ describe('script', () => {
     });
 
     it('work for complex objects', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
 
         await page.setContent(html`<script src="./scripts.js"></script> `);
@@ -85,7 +85,7 @@ describe('script', () => {
     });
 
     it('work for async functions', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
 
         await page.setContent(html`<script src="./scripts.js"></script> `);
@@ -108,7 +108,7 @@ describe('script', () => {
     });
 
     it('work with one argument', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
 
         await page.setContent(html`<button id="test">test</button>`);
@@ -133,7 +133,7 @@ describe('script', () => {
     });
 
     it('work with multiple args', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
 
         await page.setContent(html`<button id="test">test</button>`);
@@ -164,7 +164,7 @@ describe('script', () => {
       );
       server.addHtmlRoute('/main', html`<iframe src="/iframe"></iframe>`);
 
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
         await page.goto(server.getRoute('/main'));
         await context.createTextSnapshot();
