+++ tests/tools/snapshot.test.ts
@@ -8,12 +8,12 @@ import assert from 'node:assert';
 import {describe, it} from 'node:test';
 
 import {takeSnapshot, waitFor} from '../../src/tools/snapshot.js';
-import {html, withBrowser} from '../utils.js';
+import {html, withMcpContext} from '../utils.js';
 
 describe('snapshot', () => {
   describe('browser_snapshot', () => {
     it('includes a snapshot', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         await takeSnapshot.handler({params: {}}, response, context);
         assert.ok(response.includeSnapshot);
       });
@@ -21,7 +21,7 @@ describe('snapshot', () => {
   });
   describe('browser_wait_for', () => {
     it('should work', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
 
         await page.setContent(
@@ -45,7 +45,7 @@ describe('snapshot', () => {
       });
     });
     it('should work with element that show up later', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
 
         const handlePromise = waitFor.handler(
@@ -72,7 +72,7 @@ describe('snapshot', () => {
       });
     });
     it('should work with aria elements', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
 
         await page.setContent(
@@ -98,7 +98,7 @@ describe('snapshot', () => {
     });
 
     it('should work with iframe content', async () => {
-      await withBrowser(async (response, context) => {
+      await withMcpContext(async (response, context) => {
         const page = context.getSelectedPage();
 
         await page.setContent(
